기능추가
카카오톡 로그인, 프로젝트 구조 일부 수정, 코드 일부 수정

+++ src/main/java/com/sundolls/epbackend/entity/StudyInfo.java
@@ -31,7 +31,7 @@ public class StudyInfo {
     private User user;
 
     @Column(name = "TOTAL_STUDY_TIME")
-    private Long time;
+    private Integer time;
 
     @Column(name = "CREATED_AT")
     @CreatedDate
